Extract #remove_transient_data from TruncateData
