+++ lib/tasks/data/remove_transient_data.rake
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'highline'
+require 'tasks/data/remove_transient_data'
+
+namespace :ofn do
+  namespace :data do
+    desc 'Remove transient data'
+    task remove_transient_data: :environment do
+      RemoveTransientData.new.call
+    end
+  end
+end
