+++ spec/lib/tasks/data/truncate_data_spec.rb
@@ -7,9 +7,6 @@ describe TruncateData do
   describe '#call' do
     before do
       allow(Spree::ReturnAuthorization).to receive(:delete_all)
-      allow(Spree::StateChange).to receive(:delete_all)
-      allow(Spree::LogEntry).to receive(:delete_all)
-      allow(TruncateData::Session).to receive(:delete_all)
       allow(Rails.logger).to receive(:info)
     end
 
@@ -24,30 +21,6 @@ describe TruncateData do
 
         expect(OrderCycle.all).to be_empty
       end
-
-      it 'deletes state changes older than a month' do
-        TruncateData.new.call
-
-        expect(Spree::StateChange)
-          .to have_received(:delete_all)
-          .with("created_at < '#{1.month.ago.to_date}'")
-      end
-
-      it 'deletes log entries older than a month' do
-        TruncateData.new.call
-
-        expect(Spree::LogEntry)
-          .to have_received(:delete_all)
-          .with("created_at < '#{1.month.ago.to_date}'")
-      end
-
-      it 'deletes sessions older than two weeks' do
-        TruncateData.new.call
-
-        expect(TruncateData::Session)
-          .to have_received(:delete_all)
-          .with("created_at < '#{2.weeks.ago.to_date}'")
-      end
     end
 
     context 'when months_to_keep is nil' do
@@ -60,30 +33,6 @@ describe TruncateData do
         TruncateData.new(nil).call
 
         expect(OrderCycle.all).to be_empty
-      end
-
-      it 'deletes state changes older than a month' do
-        TruncateData.new.call
-
-        expect(Spree::StateChange)
-          .to have_received(:delete_all)
-          .with("created_at < '#{1.month.ago.to_date}'")
-      end
-
-      it 'deletes log entries older than a month' do
-        TruncateData.new.call
-
-        expect(Spree::LogEntry)
-          .to have_received(:delete_all)
-          .with("created_at < '#{1.month.ago.to_date}'")
-      end
-
-      it 'deletes sessions older than two weeks' do
-        TruncateData.new.call
-
-        expect(TruncateData::Session)
-          .to have_received(:delete_all)
-          .with("created_at < '#{2.weeks.ago.to_date}'")
       end
     end
 
@@ -101,30 +50,6 @@ describe TruncateData do
         TruncateData.new(6).call
 
         expect(OrderCycle.all).to contain_exactly(recent_order_cycle)
-      end
-
-      it 'deletes state changes older than a month' do
-        TruncateData.new.call
-
-        expect(Spree::StateChange)
-          .to have_received(:delete_all)
-          .with("created_at < '#{1.month.ago.to_date}'")
-      end
-
-      it 'deletes log entries older than a month' do
-        TruncateData.new.call
-
-        expect(Spree::LogEntry)
-          .to have_received(:delete_all)
-          .with("created_at < '#{1.month.ago.to_date}'")
-      end
-
-      it 'deletes sessions older than two weeks' do
-        TruncateData.new.call
-
-        expect(TruncateData::Session)
-          .to have_received(:delete_all)
-          .with("created_at < '#{2.weeks.ago.to_date}'")
       end
     end
   end
